chore(frontend): declare module' type definition on each modules
instead of centralized in `shim.d.ts`
